Move player to own component
